Decoupled weapon utils. Moved time utils and sound utils.
