+++ src/engine/core/database/logic.ts
@@ -4,8 +4,8 @@ import { loadPortableStore, savePortableStore } from "@/engine/core/database/por
 import { registry } from "@/engine/core/database/registry";
 import { closeLoadMarker, closeSaveMarker, openLoadMarker, openSaveMarker } from "@/engine/core/database/save_markers";
 import { IBaseSchemeState, IRegistryObjectState } from "@/engine/core/database/types";
-import { readTimeFromPacket, writeTimeToPacket } from "@/engine/core/utils/game/game_time";
 import { emitSchemeEvent } from "@/engine/core/utils/scheme/scheme_event";
+import { readTimeFromPacket, writeTimeToPacket } from "@/engine/core/utils/time";
 import { NIL } from "@/engine/lib/constants/words";
 import {
   ClientObject,
