+++ src/engine/core/managers/actor/ActorInputManager.ts
@@ -15,10 +15,10 @@ import type { SurgeManager } from "@/engine/core/managers/surge/SurgeManager";
 import { killAllSurgeUnhidden } from "@/engine/core/managers/surge/utils/surge_kill";
 import { WeatherManager } from "@/engine/core/managers/weather";
 import { executeConsoleCommand, getConsoleFloatCommand } from "@/engine/core/utils/game";
-import { readTimeFromPacket, writeTimeToPacket } from "@/engine/core/utils/game/game_time";
 import { disableInfoPortion, giveInfoPortion } from "@/engine/core/utils/info_portion";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isActorInNoWeaponZone } from "@/engine/core/utils/object/object_zone";
+import { isActorInNoWeaponZone } from "@/engine/core/utils/position";
+import { readTimeFromPacket, writeTimeToPacket } from "@/engine/core/utils/time";
 import { animations, postProcessors } from "@/engine/lib/constants/animation";
 import { consoleCommands } from "@/engine/lib/constants/console_commands";
 import { infoPortions } from "@/engine/lib/constants/info_portions";
