+++ src/engine/core/managers/surge/SurgeManager.ts
@@ -22,10 +22,10 @@ import type { AnomalyZoneBinder } from "@/engine/core/objects/binders/zones";
 import { isArtefact } from "@/engine/core/utils/class_ids";
 import { isBlackScreen } from "@/engine/core/utils/game";
 import { createGameAutoSave } from "@/engine/core/utils/game/game_save";
-import { readTimeFromPacket, writeTimeToPacket } from "@/engine/core/utils/game/game_time";
 import { hasInfoPortion } from "@/engine/core/utils/info_portion";
 import { pickSectionFromCondList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { readTimeFromPacket, writeTimeToPacket } from "@/engine/core/utils/time";
 import { createVector } from "@/engine/core/utils/vector";
 import { animations, postProcessors } from "@/engine/lib/constants/animation";
 import { ACTOR_ID } from "@/engine/lib/constants/ids";
