+++ src/engine/core/objects/ai/state/weapon/ActionWeaponStrap.ts
@@ -3,7 +3,7 @@ import { action_base, LuabindClass, object } from "xray16";
 import { StalkerStateManager } from "@/engine/core/objects/ai/state/StalkerStateManager";
 import { isStrappableWeapon } from "@/engine/core/utils/class_ids";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getObjectWeaponForAnimationState } from "@/engine/core/utils/object/object_weapon";
+import { getObjectWeaponForAnimationState } from "@/engine/core/utils/weapon";
 import { ClientObject, Optional } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
