+++ src/engine/core/objects/ai/state/weapon/ActionWeaponUnstrap.ts
@@ -2,10 +2,7 @@ import { action_base, LuabindClass } from "xray16";
 
 import { StalkerStateManager } from "@/engine/core/objects/ai/state/StalkerStateManager";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import {
-  getObjectWeaponForAnimationState,
-  getWeaponActionForAnimationState,
-} from "@/engine/core/utils/object/object_weapon";
+import { getObjectWeaponForAnimationState, getWeaponActionForAnimationState } from "@/engine/core/utils/weapon";
 
 const logger: LuaLogger = new LuaLogger($filename);
