+++ src/engine/core/objects/ai/state/weapon/EvaluatorWeaponLocked.ts
@@ -2,7 +2,7 @@ import { LuabindClass, property_evaluator } from "xray16";
 
 import { StalkerStateManager } from "@/engine/core/objects/ai/state/StalkerStateManager";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isObjectWeaponLocked } from "@/engine/core/utils/object/object_weapon";
+import { isObjectWeaponLocked } from "@/engine/core/utils/weapon";
 
 const logger: LuaLogger = new LuaLogger($filename);
