+++ src/engine/core/schemes/restrictor/sr_timer/TimerManager.ts
@@ -3,10 +3,10 @@ import { CUIGameCustom, get_hud, time_global } from "xray16";
 import { registry } from "@/engine/core/database";
 import { AbstractSchemeManager } from "@/engine/core/objects/ai/scheme";
 import { ETimerType, ISchemeTimerState } from "@/engine/core/schemes/restrictor/sr_timer/ISchemeTimerState";
-import { globalTimeToString } from "@/engine/core/utils/game/game_time";
 import { pickSectionFromCondList } from "@/engine/core/utils/ini/ini_config";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { switchObjectSchemeToSection, trySwitchToAnotherSection } from "@/engine/core/utils/scheme/scheme_switch";
+import { globalTimeToString } from "@/engine/core/utils/time";
 import { Optional, TSection, TTimestamp } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
