+++ src/engine/core/utils/game/game_save.ts
@@ -4,8 +4,8 @@ import { registry } from "@/engine/core/database/registry";
 import { assert } from "@/engine/core/utils/assertion";
 import { loadObjectFromFile, saveObjectToFile } from "@/engine/core/utils/fs";
 import { executeConsoleCommand } from "@/engine/core/utils/game/game_console";
-import { gameTimeToString } from "@/engine/core/utils/game/game_time";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { gameTimeToString } from "@/engine/core/utils/time";
 import { gameConfig } from "@/engine/lib/configs/GameConfig";
 import { consoleCommands } from "@/engine/lib/constants/console_commands";
 import { TGameDifficulty } from "@/engine/lib/constants/game_difficulties";
