+++ src/engine/core/utils/game/index.ts
@@ -1,5 +1,4 @@
 export * from "@/engine/core/utils/game/game_console";
 export * from "@/engine/core/utils/game/game_save";
-export * from "@/engine/core/utils/game/game_time";
 export * from "@/engine/core/utils/game/game_check";
 export * from "@/engine/core/utils/game/game_wait";
