--- src/engine/core/utils/object/object_zone.test.ts
@@ -1,5 +1,0 @@
-import { describe, it } from "@jest/globals";
-
-describe("object_zone utils", () => {
-  it.todo("isActorInNoWeaponZone should correctly check if actor is in no weapon zone");
-});
