--- src/engine/core/utils/object/object_zone.ts
@@ -1,16 +1,0 @@
-import { registry } from "@/engine/core/database";
-
-/**
- * todo;
- * todo;
- * todo;
- */
-export function isActorInNoWeaponZone(): boolean {
-  for (const [, isActive] of registry.noWeaponZones) {
-    if (isActive) {
-      return true;
-    }
-  }
-
-  return false;
-}
