+++ src/engine/core/utils/sound.ts
@@ -2,7 +2,7 @@ import { bit_and, snd_type } from "xray16";
 
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { ESoundType } from "@/engine/lib/constants/sound";
-import { TSoundType } from "@/engine/lib/types";
+import { ClientObject, TSoundType } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
 
@@ -69,4 +69,16 @@ export function mapSoundMaskToSoundType(soundMask: TSoundType): ESoundType {
  */
 export function isSoundType(heard: TSoundType, expected: TSoundType): boolean {
   return bit_and(heard, expected) === expected;
+}
+
+/**
+ * Stop playing sound for client game object.
+ *
+ * @param object - target object to stop playing
+ */
+export function stopPlayingObjectSound(object: ClientObject): void {
+  if (object.alive()) {
+    object.set_sound_mask(-1);
+    object.set_sound_mask(0);
+  }
 }
