+++ src/engine/core/utils/weapon.ts
